Make sure we still support ERB templates when provided
`apache::mod::php` allows to pass an ERB template, switching the default
template to EPP will require us to change the default value of the
`template` parameter which is generally a breaking change.

Users who rely on this parameter to provide a custom template are
currently using an ERB template, so we must preserve the legacy behavior
for them, and detect if the template should be processed as ERB or EPP.

For this purpose, we check the file extension in a conservative way (any
template whose filename does not end with `.epp` is assumbed to be an
ERB template).

As a result, this change is backwards-compatible for end-users.